Core: x64 cosmetics
git-svn-id: svn://ultimatepp.org/upp/trunk@5199 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/InetUtil.cpp
@@ -58,7 +58,7 @@ static const char hex_digits[] = "0123456789ABCDEF";
 String UrlEncode(const char *p, const char *e)
 {
 	StringBuffer out;
-	out.Reserve(e - p);
+	out.Reserve((int)(e - p));
 	for(; p < e; p++)
 	{
 		const char *b = p;
